backend/requirements.txt
autopep8==1.5.4
coverage==5.3
flake8==3.8.4
Flask==1.1.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytest==6.1.1
pytest-cov==2.10.1
webargs==6.1.1